+++ src/item.cpp
@@ -107,6 +107,7 @@ void BarrelReceiver::onCollide(PhysicsBody& other)
     {
         other.setDead(true);
         activated = true;
+        Globals::Game.terrain.emitSignal(signal,this);
         if (onTrigger.get())
         {
             (*onTrigger)(*this);