+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * TODO: reverse gravity is cool, planets need some kind of particle effect pushing away to very obviously show where the field is strong, its not intuitive like normal gravity
 
 ## NEW RELEASE
+* Added the `signals` system, which allows for level configs to trigger C++ code via strings
+* Fixed a bug causing parallel horizontal lines to count as intersecting when they were not on the same `y`
+* Added rectangular moving terrain (their collision is ass around the corners)
+* Removed the `pointInRectAngle` function (literally never used and has a confusing name)
 
 ## 4/11/2026 
 * Removed Barrel's Factory to prevent it from being saved in EDITOR mode.